+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Other Changes:
 - Added Server-side Rendering Support.
 - Fixed a reference cycle between `Style` and `StyleManager`.
+- Added `StyleManager::new()` to create a style manager with default configuration.
 
 ## v0.11.0